Simplifies GTFS export triggering
* Triggering of GTFS merged exports is now handled by sending a single message over pubsub regardless of whether we are exporting basic or extended gtfs - both are always merged in damu. This commit cleans up code of this distinction as it is no longer needed.
* This commit also removes the possibility of configuring whether to include shapes or not in export of gtfs basic merge. If we are going to preserve this config property going forward, it will likely be implemented in damu instead.
